wishlist, car cards - code refactor 🚀
