+++ blocks/car-card/car-card.css
@@ -1,154 +1,153 @@
-.section.car-details {
-  .car-cards-wrapper,
-  .recently-viewed-cards-wrapper {
-    list-style-type: none;
-    padding: 0;
+.car-cards-wrapper,
+.wishlist-cards-wrapper,
+.recently-viewed-cards-wrapper {
+  list-style-type: none;
+  padding: 0;
+  margin: 0;
+  display: grid;
+  width: 100%;
+  gap: 1rem;
+  grid-template-columns: 1fr;
+
+  p,
+  h1,
+  h2,
+  h3,
+  h4,
+  h5,
+  h6,
+  a {
     margin: 0;
-    display: grid;
-    width: 100%;
+  }
+
+  .car-detail-container {
+    border: 1px solid var(--border-color);
+    border-radius: 0.5rem;
+    padding: var(--spacing-xs) var(--spacing-xs) var(--spacing-s) var(--spacing-xs);
     gap: 1rem;
-    grid-template-columns: 1fr;
+    display: flex;
+    flex-direction: column;
+    font-size: var(--font-size-xs);
+    line-height: 1.5rem;
+    box-shadow: var(--box-shadow);
+    box-sizing: border-box;
 
-    p,
-    h1,
-    h2,
-    h3,
-    h4,
-    h5,
-    h6,
-    a {
-      margin: 0;
-    }
-
-    .car-detail-container {
-      border: 1px solid var(--border-color);
-      border-radius: 0.5rem;
-      padding: var(--spacing-xs) var(--spacing-xs) var(--spacing-s) var(--spacing-xs);
-      gap: 1rem;
+    .container {
       display: flex;
       flex-direction: column;
-      font-size: var(--font-size-xs);
-      line-height: 1.5rem;
-      box-shadow: var(--box-shadow);
-      box-sizing: border-box;
+      gap: 1rem;
 
-      .container {
+      h3 {
+        font-size: var(--font-size-l);
+        line-height: 2rem;
+        font-weight: 600;
+      }
+
+      .heart-container {
         display: flex;
-        flex-direction: column;
+        justify-content: space-between;
         gap: 1rem;
 
-        h3 {
-          font-size: var(--font-size-l);
-          line-height: 2rem;
-          font-weight: 600;
-        }
-
-        .heart-container {
-          display: flex;
-          justify-content: space-between;
-          gap: 1rem;
-
-          .car-card-heart {
-            cursor: pointer;
-            width: 1.5rem;
-            height: 1.5rem;
-
-            &:hover {
-              transform: scale(1.25);
-              transition: transform 0.2s;
-            }
-          }
-        }
-
-        .car-image-container {
-          position: relative;
-          background: var(--bg-color-neutral);
-          display: flex;
-          align-items: center;
-          justify-content: center;
-          width: 100%;
-          height: 153px;
-          margin: 0 auto;
-
-          .car-badge-container {
-            display: flex;
-            justify-content: space-between;
-            position: absolute;
-            top: 0;
-            left: 0;
-            width: 100%;
-
-            .oly-deals {
-              width: 72px;
-              height: 48px;
-              box-shadow: var(--box-shadow-car-card);
-              border-radius: 0 0 100px 100px;
-              background: var(--color-light-pink);
-              text-align: center;
-              font-size: var(--font-size-xs);
-              line-height: 1.125rem;
-              font-weight: 800;
-              color: var(--color-black);
-              position: absolute;
-              left: var(--spacing-xxxs);
-            }
-
-            .ev-icon {
-              width: 65px;
-              height: 20px;
-              position: absolute;
-              top: var(--spacing-xxxs);
-              right: var(--spacing-xxxs);
-            }
-          }
-
-          picture img {
-            width: 264px;
-            aspect-ratio: 2.44;
-            display: block;
-          }
-        }
-
-        .car-card-description {
-          color: var(--color-black);
-        }
-      }
-
-      .car-card-price-container {
-        .car-card-price {
-          font-size: var(--font-size-lg-alt-s);
-          line-height: 2.5rem;
-          font-weight: 600;
-        }
-
-        .car-card-price-title,
-        .car-card-price-weekly,
-        .car-card-info {
-          font-size: var(--font-size-xs-alt);
-          color: var(--color-black);
-        }
-
-        .car-card-price-weekly,
-        .car-card-info {
-          font-weight: 700;
-        }
-      }
-
-      .car-card-button-container {
-        a {
-          margin: 0;
-          width: 100%;
-          display: flex;
-          justify-content: center;
-          gap: 1rem;
-        }
-
-        a::after {
-          content: url('../../icons/chevron-right-white.svg');
+        .car-card-heart {
+          cursor: pointer;
           width: 1.5rem;
           height: 1.5rem;
-          padding-top: 2px;
+
+          &:hover {
+            transform: scale(1.25);
+            transition: transform 0.2s;
+          }
         }
+      }
+
+      .car-image-container {
+        position: relative;
+        background: var(--bg-color-neutral);
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        width: 100%;
+        height: 153px;
+        margin: 0 auto;
+
+        .car-badge-container {
+          display: flex;
+          justify-content: space-between;
+          position: absolute;
+          top: 0;
+          left: 0;
+          width: 100%;
+
+          .oly-deals {
+            width: 72px;
+            height: 48px;
+            box-shadow: var(--box-shadow-car-card);
+            border-radius: 0 0 100px 100px;
+            background: var(--color-light-pink);
+            text-align: center;
+            font-size: var(--font-size-xs);
+            line-height: 1.125rem;
+            font-weight: 800;
+            color: var(--color-black);
+            position: absolute;
+            left: var(--spacing-xxxs);
+          }
+
+          .ev-icon {
+            width: 65px;
+            height: 20px;
+            position: absolute;
+            top: var(--spacing-xxxs);
+            right: var(--spacing-xxxs);
+          }
+        }
+
+        picture img {
+          width: 264px;
+          aspect-ratio: 2.44;
+          display: block;
+        }
+      }
+
+      .car-card-description {
+        color: var(--color-black);
+      }
+    }
+
+    .car-card-price-container {
+      .car-card-price {
+        font-size: var(--font-size-lg-alt-s);
+        line-height: 2.5rem;
+        font-weight: 600;
+      }
+
+      .car-card-price-title,
+      .car-card-price-weekly,
+      .car-card-info {
+        font-size: var(--font-size-xs-alt);
+        color: var(--color-black);
+      }
+
+      .car-card-price-weekly,
+      .car-card-info {
+        font-weight: 700;
+      }
+    }
+
+    .car-card-button-container {
+      a {
+        margin: 0;
+        width: 100%;
+        display: flex;
+        justify-content: center;
+        gap: 1rem;
+      }
+
+      a::after {
+        content: url('../../icons/chevron-right-white.svg');
+        width: 1.5rem;
+        height: 1.5rem;
+        padding-top: 2px;
       }
     }
   }
@@ -157,22 +156,24 @@
 @media (width >= 768px) {
   .section.car-details {
     padding: var(--spacing-s) var(--spacing-l-alt);
+  }
 
-    .car-cards-wrapper,
-    .recently-viewed-cards-wrapper {
-      grid-template-columns: repeat(2, 1fr);
-    }
+  .car-cards-wrapper,
+  .wishlist-cards-wrapper,
+  .recently-viewed-cards-wrapper {
+    grid-template-columns: repeat(2, 1fr);
   }
 }
 
 @media (width >=1024px) {
   .section.car-details {
     padding: var(--spacing-l-alt) var(--spacing-xxl-l);
+  }
 
-    .car-cards-wrapper,
-    .recently-viewed-cards-wrapper {
-      grid-template-columns: repeat(3, 1fr);
-    }
+  .car-cards-wrapper,
+  .wishlist-cards-wrapper,
+  .recently-viewed-cards-wrapper {
+    grid-template-columns: repeat(3, 1fr);
   }
 }
